refactor(harness): split the misnamed environment module into catalogue and simulator, and keep prompt text out of code
