+++ src/fi/alk/harness/run/live.py
@@ -23,7 +23,8 @@ from dataclasses import dataclass, field
 from pathlib import Path
 
 from ..checks import Outcome, run_check
-from ..environment import fill, load_catalogue, load_simulator_prompt
+from ..catalogue import load_catalogue
+from ..simulator import fill, load_simulator_prompt
 from ..scenario import Scenario
 from ..world.runtime import GeneratedWorld
 from ..folder import apply_setup, check_ready
@@ -162,19 +163,16 @@ def grade(scenario: Scenario, world: GeneratedWorld, world_root: Path) -> LiveRu
 
 def instruction_for(scenario: Scenario, world_root: Path) -> str:
     """What the simulated person is told, from the prompt the environment step wrote."""
-    from ..environment import with_rules
-
     written = load_simulator_prompt(world_root)
     if not written:
-        # Even with no prompt written for this agent, the caller is still on a phone.
-        return with_rules(scenario.instruction, spoken=True)
+        return scenario.instruction
     filled, missing = fill(written, scenario.slots())
     if missing:
         raise RuntimeError(
             f"the simulator prompt asks for {', '.join(missing)}, which {scenario.name} does "
             "not supply. An unfilled slot reaches the caller verbatim."
         )
-    return with_rules(filled, spoken=True)
+    return filled
 
 
 def wire(
